Update the readme, change the script name, and tweak it
